# Conflicts:
#	src/main/java/xyz/imaf6971/volgaitfinal/model/User.java
#	src/main/java/xyz/imaf6971/volgaitfinal/service/impl/UserServiceImpl.java

+++ src/main/java/xyz/imaf6971/volgaitfinal/service/impl/UserServiceImpl.java
@@ -48,7 +48,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(String username) {
-        repository.deleteByUsername(username);
+        if (getCurrentUser().isAdmin()) {
+            repository.deleteByUsername(username);
+        }
     }
 
     public void restrictAdvertisementPublishing(String username) {
